Added the aritcle page for error in the slugs part
